Move capacity mesh outline fixture next to tests
